requirements.txt
fastapi
uvicorn
easyocr
deep-translator
pyttsx3
websockets
python-multipart
pillow
numpy
